+++ src/WinForge.App/Views/StorageView.xaml.cs
@@ -1,0 +1,8 @@
+using System.Windows.Controls;
+
+namespace WinForge.App.Views;
+
+public partial class StorageView : UserControl
+{
+    public StorageView() => InitializeComponent();
+}